Improved readability of info subcommands.
Added description and ex fields to Info struct in Geolocation.

Added ToDescription method in to geolocation.Info. This prints the
description and example fields (if applicable) for all fields in the Info
struct. This method also adds some styling for easier readbility in
monochrome terminals. This also makes updating the info command automatic
as more fields are added to the Info struct.

Adds fatih/color package for ANSI coloring.